requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
aiohappyeyeballs=2.4.2=pypi_0
aiohttp=3.10.8=pypi_0
aiosignal=1.3.1=pypi_0
annotated-types=0.7.0=pypi_0
anyio=4.6.0=pypi_0
attrs=24.2.0=pypi_0
backoff=2.2.1=pypi_0
beautifulsoup4=4.12.3=pypi_0
bzip2=1.0.8=h6c40b1e_6
ca-certificates=2024.7.2=hecd8cb5_0
certifi=2024.8.30=pypi_0
cffi=1.17.1=pypi_0
chardet=5.2.0=pypi_0
charset-normalizer=3.3.2=pypi_0
click=8.1.7=pypi_0
cryptography=43.0.1=pypi_0
dataclasses-json=0.6.7=pypi_0
deepdiff=8.0.1=pypi_0
distro=1.9.0=pypi_0
emoji=2.13.2=pypi_0
expat=2.6.3=h6d0c2b6_0
filetype=1.2.0=pypi_0
frozenlist=1.4.1=pypi_0
greenlet=3.1.1=pypi_0
h11=0.14.0=pypi_0
httpcore=1.0.5=pypi_0
httpx=0.27.2=pypi_0
idna=3.10=pypi_0
jiter=0.5.0=pypi_0
joblib=1.4.2=pypi_0
jsonpatch=1.33=pypi_0
jsonpath-python=1.0.6=pypi_0
jsonpointer=3.0.0=pypi_0
langchain=0.3.1=pypi_0
langchain-community=0.3.1=pypi_0
langchain-core=0.3.6=pypi_0
langchain-text-splitters=0.3.0=pypi_0
langdetect=1.0.9=pypi_0
langsmith=0.1.129=pypi_0
libcxx=14.0.6=h9765a3e_0
libffi=3.4.4=hecd8cb5_1
lxml=5.3.0=pypi_0
markdown=3.7=pypi_0
marshmallow=3.22.0=pypi_0
multidict=6.1.0=pypi_0
mypy-extensions=1.0.0=pypi_0
ncurses=6.4=hcec6c5f_0
nest-asyncio=1.6.0=pypi_0
nltk=3.9.1=pypi_0
numpy=1.26.4=pypi_0
olefile=0.47=pypi_0
openai=1.50.2=pypi_0
openssl=3.0.15=h46256e1_0
orderly-set=5.2.2=pypi_0
orjson=3.10.7=pypi_0
packaging=24.1=pypi_0
pip=24.2=py312hecd8cb5_0
psutil=6.0.0=pypi_0
pycparser=2.22=pypi_0
pydantic=2.9.2=pypi_0
pydantic-core=2.23.4=pypi_0
pydantic-settings=2.5.2=pypi_0
pypdf=5.0.1=pypi_0
python=3.12.5=hcd54a6c_1
python-dateutil=2.9.0.post0=pypi_0
python-dotenv=1.0.1=pypi_0
python-iso639=2024.4.27=pypi_0
python-magic=0.4.27=pypi_0
python-oxmsg=0.0.1=pypi_0
pyyaml=6.0.2=pypi_0
rapidfuzz=3.10.0=pypi_0
readline=8.2=hca72f7f_0
regex=2024.9.11=pypi_0
requests=2.32.3=pypi_0
requests-toolbelt=1.0.0=pypi_0
setuptools=75.1.0=py312hecd8cb5_0
six=1.16.0=pypi_0
sniffio=1.3.1=pypi_0
soupsieve=2.6=pypi_0
sqlalchemy=2.0.35=pypi_0
sqlite=3.45.3=h6c40b1e_0
tabulate=0.9.0=pypi_0
tenacity=8.5.0=pypi_0
tk=8.6.14=h4d00af3_0
tqdm=4.66.5=pypi_0
typing-extensions=4.12.2=pypi_0
typing-inspect=0.9.0=pypi_0
tzdata=2024a=h04d1e81_0
unstructured=0.15.13=pypi_0
unstructured-client=0.25.9=pypi_0
urllib3=2.2.3=pypi_0
wheel=0.44.0=py312hecd8cb5_0
wrapt=1.16.0=pypi_0
xz=5.4.6=h6c40b1e_1
yarl=1.13.1=pypi_0
zlib=1.2.13=h4b97444_1
